refactor: rename vk_loader.cpp/.h
