+++ src/graphics/CMakeLists.txt
@@ -1,11 +1,11 @@
 target_sources(${PROJECT_NAME}
         PRIVATE
         vulkan/MeshNode.cpp
+        vulkan/RenderableGLTF.cpp
         vulkan/vk_descriptors.cpp
         vulkan/vk_engine.cpp
         vulkan/vk_images.cpp
         vulkan/vk_initializers.cpp
-        vulkan/vk_loader.cpp
         vulkan/vk_pipelines.cpp
         Graphics.cpp
 )
